// src/model/policy-schema/policy.definitions.ts
export enum HostNameTypeEnum {
  Explicit = "explicit",
  Wildcard = "wildcard",
}
// ...
export enum ParameterLocation {
  Any = "any",
  Cookie = "cookie",
  FormData = "form-data",
  Header = "header",
  Path = "path",
  Query = "query",
}
// ...
export interface Header {
  allowRepeatedOccurrences?: boolean;
  base64Decoding?: boolean;
  checkSignatures?: boolean;
  htmlNormalization?: boolean;
  mandatory?: boolean;
  maskValueInLogs?: boolean;
  name: string;
  normalizationViolations?: boolean;
  percentDecoding?: boolean;
  signatureOverrides?: HeaderSignatureOverride[];
  type?: HostNameTypeEnum;
  urlNormalization?: boolean;
  wildcardOrder?: number;
}

export interface HeaderSignatureOverride {
  enabled?: boolean;
  name?: string;
  signatureId?: number;
  tag?: string;
}